requirements.txt
tensorflow==2.3.1
pandas==1.1.5
dlib==19.21.1
matplotlib==3.3.3
numpy==1.18.5
opencv_python==4.4.0.46
Pillow==8.2.0
scikit_learn==0.24.2
